Dimecoin 2.5.5.2 - headers-sync stall fix and 15 low-severity fixes

+++ src/coins.h
@@ -167,7 +167,7 @@ public:
 
     virtual bool GetKey(COutPoint &key) const = 0;
     virtual bool GetValue(Coin &coin) const = 0;
-    virtual unsigned int GetValueSize() const = 0;
+    virtual size_t GetValueSize() const = 0;
 
     virtual bool Valid() const = 0;
     virtual void Next() = 0;
